STAR-543 Port remaining guardrails and sync with cndb (#180)
(cherry picked from commit a83e0e87c5521ec24e76de57db03dc0ed093c73f)
(cherry picked from commit b90b43306e89172ae7882cb041157abf7e30740e)
(cherry picked from commit 36600dfeadaad6affe2b264e3e8f07853b7c94df)
(cherry picked from commit abc39d60b9b82178d86cfc86c880ea7978835fe0)

+++ src/java/org/apache/cassandra/db/filter/TombstoneOverwhelmingException.java
@@ -26,7 +26,7 @@ import org.apache.cassandra.db.marshal.*;
 
 public class TombstoneOverwhelmingException extends RuntimeException
 {
-    public TombstoneOverwhelmingException(int numTombstones, String query, TableMetadata metadata, DecoratedKey lastPartitionKey, ClusteringPrefix<?> lastClustering)
+    public TombstoneOverwhelmingException(long numTombstones, String query, TableMetadata metadata, DecoratedKey lastPartitionKey, ClusteringPrefix<?> lastClustering)
     {
         super(String.format("Scanned over %d tombstones during query '%s' (last scanned row token was %s and partion key was (%s)); query aborted",
                             numTombstones, query, lastPartitionKey.getToken(), makePKString(metadata, lastPartitionKey.getKey(), lastClustering)));
